Feat : added sync mattermost user info with a mattermost info table in db
