+++ src/lib/mattermost/mattermost.ts
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import config from '@config';
+import { getMattermostConfig } from '.';
 
 export interface MattermostUser {
   id: string,
@@ -18,21 +19,6 @@ export interface MattermostUser {
   mfa_active: boolean,
   last_activity_at: string,
 }
-
-
-const getMattermostConfig = () => {
-  if (!config.mattermostBotToken) {
-    const errorMessage =
-      'Unable to launch mattermost api calls without env var mattermostBotToken';
-    console.error(errorMessage);
-    throw new Error(errorMessage);
-  }
-  return {
-    headers: {
-      Authorization: `Bearer ${config.mattermostBotToken}`,
-    },
-  };
-};
 
 export async function getUserWithParams(params = {}, i = 0) {
   const mattermostUsers = await axios
