+++ src/schedulers/mattermostScheduler/inviteUserToTeamByEmail.ts
@@ -1,0 +1,17 @@
+import config from "@/config";
+import { MemberWithPrimaryEmail } from "@/models/member";
+import * as mattermost from '@/lib/mattermost';
+import { getActiveGithubUsersUnregisteredOnMattermost } from ".";
+
+export async function inviteUsersToTeamByEmail() {
+    const activeGithubUsersNotInCommunityTeam: MemberWithPrimaryEmail[] =
+      await getActiveGithubUsersUnregisteredOnMattermost();
+    const results = await mattermost.inviteUsersToTeamByEmail(
+      activeGithubUsersNotInCommunityTeam
+        .map((user) => user.primary_email)
+        .slice(0, 19),
+      config.mattermostTeamId
+    );
+    return results;
+  }
+
