+++ src/scripts/mattermost/runSyncMattermostUserWithMattermostMemberInfosTable.ts
@@ -1,0 +1,5 @@
+import { syncMattermostUserWithMattermostMemberInfosTable } from "@/schedulers/mattermostScheduler/syncMattermostUserWithMattermostMemberInfosTable"
+
+syncMattermostUserWithMattermostMemberInfosTable().then(() => {
+    console.log('Sync syncMattermostUserWithMattermostMemberInfosTable done')
+})
